+++ p2panda-rs/src/wasm/tests/mod.rs
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+//! Tests for `wasm` module in `p2panda_rs`.
+mod entry;
+mod key_pair;
+mod operation;